+++ src/Code/NSShadow+Scaling.m
@@ -50,8 +50,8 @@
 {
 	NSSize	offset;
 	
-	offset.width = dist * cosf( radians );
-	offset.height = dist * sinf( radians );
+	offset.width = dist * cos( radians );
+	offset.height = dist * sin( radians );
 	
 	[self setShadowOffset:offset];
 }
@@ -66,7 +66,7 @@
 - (CGFloat)		shadowAngle
 {
 	NSSize offset = [self shadowOffset];
-	return atan2f( offset.height, offset.width );
+	return atan2( offset.height, offset.width );
 }
 
 
@@ -87,8 +87,8 @@
 {
 	NSSize	offset;
 	
-	offset.width = [self distance] * cosf( radians );
-	offset.height = [self distance] * sinf( radians );
+	offset.width = [self distance] * cos( radians );
+	offset.height = [self distance] * sin( radians );
 	
 	[self setShadowOffset:offset];
 }
@@ -99,7 +99,7 @@
 	if( degrees < 0 )
 		degrees += 360;
 	
-	degrees = fmodf( degrees, 360.0 );
+	degrees = fmod( degrees, 360.0 );
 	
 	[self setAngle:DEGREES_TO_RADIANS(degrees)];
 }
@@ -108,7 +108,7 @@
 - (CGFloat)		angle
 {
 	NSSize offset = [self shadowOffset];
-	return atan2f( offset.height, offset.width );
+	return atan2( offset.height, offset.width );
 }
 
 
@@ -129,8 +129,8 @@
 	NSSize	offset;
 	CGFloat	radians = [self angle];
 	
-	offset.width = distance * cosf( radians );
-	offset.height = distance * sinf( radians );
+	offset.width = distance * cos( radians );
+	offset.height = distance * sin( radians );
 	
 	[self setShadowOffset:offset];
 }
@@ -139,7 +139,7 @@
 - (CGFloat)		distance
 {
 	NSSize offset = [self shadowOffset];
-	return hypotf( offset.width, offset.height );
+	return hypot( offset.width, offset.height );
 }
